backend/requirements.txt
fastapi==0.135.1
uvicorn==0.27.1
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-multipart==0.0.22
python-dotenv==1.0.1
chromadb==1.5.5
pydantic==2.12.5
pydantic-settings==2.13.1
requests==2.32.5
PyPDF2==3.0.1
PyJWT==2.12.0
passlib[bcrypt]==1.7.4
